requirements.txt
torch==1.11.0
torchvision==0.14.0
streamlit>=1.0.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
numpy>=1.19.0
pillow>=8.0.0
pandas>=1.2.0
tqdm>=4.50.0
scipy>=1.6.0
opencv-python>=4.5.0
requests>=2.25.0 